RDKEMW-4075: Enable TTML override on runtime instead of buildtime
Certain product variants have had a patch for subttxrend-ttml for
forcing center-alignment of TTML texts. Moving towards a single source
for all variants, we have to move this into runtime.
The solution here is to pass a list of TTML style attributes to the
TtmlController. They will make it into the DocumentInstance and be
applied to all texts displayed.